Initial implementation of world generation.
This is a very basic implementation, which is made worse by the fact that the only buildings that can be placed currently are skyscrapers!  But when more tiles are added to the OthersTileSet (smaller buildings, paved areas, grass areas, etc.) it should start to look somewhat better.  Then when we add better road generation, weight maps, and so on - better still.  It is, at least, *something*.