Implemented simple destructible objects.
Created a SleepyRigidBody script can be added to anything with a RigidBody.  It just sleeps the RigidBody on Awake so that it's ignored for physics calculations.  When anything collides with the object it wakes up and then reacts to physics.

This is super simple and not ideal for a lot of things, but it's a very quick-and-easy win that looks good in-game.

Applied this to the trees which are in the park tile.