Created Destructible Toggle script
This sets the object to be kinematic (static) on awake, then on collision if the impulse is above the specified threshold it will switch the object to not be kinematic, thus enabling it to react to physics.

Switched the trees to use this - so now they need a bit of force to knock over and act as a bit of a speed bump (a side-effect of how this is done, not something I have direct control of yet).