Vehicle Path Controller implemented.
The core of the vehicle path controller is now implemented and working - a vehicle controller will continuously drive around paths in tiles until it can't find one and will then despawn.

Still need to implement all the paths for the road tiles, improve the vehicle spawner, and improve the vehicle controller so it will properly take into account distance not percentage of path, accelerate/brake, slow/stop when obstructed, etc.  But we should be able to get something passable pretty quickly now.