World generation now supports random seeds.
I added a seed property to the world which if is set to a value less than 0, which it is by default, then the seed is randomly determined.  This seed is then used to compute the chunk seed - so for any given world seed the chunks will all be predictable.