Vehicle spawn on world initialisation fixed.
Road tiles to spawn on are no longer picked completely at random - when a vehicle is spawned on a road tile that road tile is removed from the pool of tiles that can be spawned on in this initial process, preventing vehicles from spawning inside each other.

This does not fix the general re-spawning issue, which is the next thing to fix!