Fixed vehicle despawning/respawning on collision.
After collision we were using the normal cleaner script, which would correctly remove the vehicle but not call the spawner to spawn a new one.

So vehicles have their own timer and when it hits the limit they call the spawner to remove the vehicle, which also spawns a new one in.